compiles on AIX
git-svn-id: https://gforge.info.ucl.ac.be/svn/mozart@3398 ada56829-ad1f-0410-b00f-83cda6628aec

+++ platform/emulator/print.cc
@@ -658,7 +658,7 @@ PRINT(HeapChunk)
 {
   CHECKDEPTH;
   stream << indent(offset)
-	 << "heap chunk: " << chunk_size << " bytes at " << this << '.';
+	 << "heap chunk: " << (int) chunk_size << " bytes at " << this << '.';
 /*
   char * data = (char *) chunk_data;
   for (int i = 0; i < chunk_size; i += 1)
@@ -672,7 +672,7 @@ PRINTLONG(HeapChunk)
 {
   CHECKDEPTHLONG;
   stream << indent(offset)
-	 << "heap chunk: " << chunk_size << " bytes at " << this << '.';
+	 << "heap chunk: " << (int) chunk_size << " bytes at " << this << '.';
 }
